+ rc: report: add columnated versions of report rclines to ease edit
taskwarrior doesn't accept spaces within the fields, so we'll just add a
second [identical] definition in comments but with spaces.  this means
we can yank/put the commented versions (to make a copy) and edit them,
removing tabs at the end with a substitution command.  this will make it
far easier to edit